Time: 5 ms (97.67%), Space: 11.5 MB (13.32%) - LeetHub

+++ 3084-count-substrings-starting-and-ending-with-given-character/3084-count-substrings-starting-and-ending-with-given-character.cpp
@@ -6,13 +6,13 @@ public:
         ios_base::sync_with_stdio(0);
         cin.tie(0);
         cout.tie(0);
-        long long cnt = 0;
+        int cnt = 0;
         for(char &itr : s)
         {
             if(itr == c)
                 cnt++;
         }
-        long long ans = (cnt*(cnt+1))/2;
+        long long ans = ((long long)cnt*(cnt+1))/2;
         return ans;
     }
 };
